+++ src/components/selectors.ts
@@ -22,7 +22,7 @@ export const sniffSelectorType = <RootState, S>(selector?: Selector<RootState, S
 
 /** @hidden */
 export const resolver = <RootState, S>(selector?: Selector<RootState, S>) => ({
-  property: (state: any) => state[selector as PropertySelector],
+  property: (state: any) => state ? state[selector as PropertySelector] : undefined,
   path: (state: RootState) => getIn(state, selector as PathSelector),
   function: selector as FunctionSelector<RootState, S>,
   nil: (state: RootState) => state,